Update market_api.py add ssg
