+++ flask/search_replace.json
@@ -14,5 +14,9 @@
   {
     "input": "메론",
     "output": "멜론"
+  },
+  {
+    "input": "달걀",
+    "output": "계란"
   }
 ]
